Compute z-scores for the top features in each topic within a MuData object.

This function calculates z-scores for the top features associated with each topic in the specified modality or across all modalities of a Seurat object. Z-scores are computed using normalized and log-transformed raw count data, allowing for a standardized comparison of feature expression levels relative to their mean and standard deviation across all cells. Computed z-scores are capped within a specified threshold range to limit extreme values.

zscores(object, mod = NULL, n_top = 10, thr = 5)

Arguments

object

A Seurat object containing multimodal single-cell data.

mod

Specific modality to compute z-scores for. If NULL, z-scores are computed for all modalities. Default is NULL.

n_top

Number of top features to select for each topic based on their importance in the topic signature. Default is 10.

thr

Threshold to cap the computed z-scores. Z-scores will be limited to the range [-thr, thr]. Default is 5.

Note

Z-scores are computed as (x - mean) / std, where x is the log- transformed expression value of a feature, mean is the mean across all cells, and std is the standard deviation across all cells.

The top n_top features for each topic are selected based on their importance in the topic signatures (highest weights).

Extreme z-scores are capped to the range [-thr, thr] to mitigate the impact of outliers.